requirements.txt
streamlit
python-dotenv
loguru
streamlit_code_editor
ipdb